+++ src/language-js/traverse/get-visitor-keys.js
@@ -1,7 +1,13 @@
 import visitorKeys from "./visitor-keys.evaluate.js";
 
 function getVisitorKeys(node) {
-  return visitorKeys[node.type]
+  const keys = visitorKeys[node.type];
+
+  if (!Array.isArray(keys) && process.env.NODE_ENV !== "production") {
+    throw new Error(`Missing visitor keys for '${node.type}'.`)
+  }
+
+  return keys
 }
 
 export default getVisitorKeys;
