+++ src/language-js/traverse/union-visitor-keys.js
@@ -1,0 +1,19 @@
+
+
+function unionVisitorKeys(visitorKeys) {
+  const result = {};
+
+  for (const keys of visitorKeys) {
+    for (const [key, value] of Object.entries(keys)) {
+      if (!result[key]) {
+        result[key] = value;
+      } else {
+        result[key] = [...new Set([...result[key], ...value])];
+      }
+    }
+  }
+
+  return result;
+}
+
+export default unionVisitorKeys
